[JBR-38-Merge-Provider] Success merged provider to one major provider. And use Layout wrapper other components.
