+++ src/components/layout/right-sidebar.tsx
@@ -1,13 +1,13 @@
-import { Jx3BoxLayoutContext } from '@components/provider/layout-provider';
-import { isApp } from '@utils/utils';
 import React, { useContext, useEffect } from 'react';
+import { Jx3BoxContext } from '@components/provider';
+import { isApp } from '@utils/utils';
 
 export interface RightSidebarProps {
   children?: any;
 }
 
 const RightSidebar: React.FC<RightSidebarProps> = props => {
-  const { rightSiderHook } = useContext(Jx3BoxLayoutContext);
+  const { rightSiderHook } = useContext(Jx3BoxContext);
   const { show, hide } = rightSiderHook;
 
   /**
